Added StbSharp as submodule
