+++ .gitmodules
@@ -7,3 +7,11 @@ url=https://github.com/MrHelmut/NVorbis.git
 [submodule "ThirdParty/SDL_GameControllerDB"]
 	path = ThirdParty/SDL_GameControllerDB
 	url = https://github.com/gabomdq/SDL_GameControllerDB.git
+[submodule "ThirdParty/StbImageWriteSharp"]
+	path = ThirdParty/StbImageWriteSharp
+	url = https://github.com/StbSharp/StbImageWriteSharp.git
+	branch = monogame
+[submodule "ThirdParty/StbImageSharp"]
+	path = ThirdParty/StbImageSharp
+	url = https://github.com/StbSharp/StbImageSharp.git
+	branch = monogame
