+++ MonoGame.Framework/Platform/Graphics/Texture2D.DirectX.cs
@@ -11,6 +11,7 @@ using SharpDX;
 using SharpDX.Direct3D11;
 using SharpDX.DXGI;
 using SharpDX.WIC;
+using StbImageSharp;
 using MapFlags = SharpDX.Direct3D11.MapFlags;
 using Resource = SharpDX.Direct3D11.Resource;
 
@@ -216,31 +217,42 @@ namespace Microsoft.Xna.Framework.Graphics
 
         private unsafe static Texture2D PlatformFromStream(GraphicsDevice graphicsDevice, Stream stream)
         {
-            int width, height, channels;
+            byte[] bytes;
+
+            // Rewind stream if it is at end
+            if (stream.CanSeek && stream.Length == stream.Position)
+            {
+                stream.Seek(0, SeekOrigin.Begin);
+            }
+
+            // Copy it's data to memory
+            // As some platforms dont provide full stream functionality and thus streams can't be read as it is
+            using (var ms = new MemoryStream())
+            {
+                stream.CopyTo(ms);
+                bytes = ms.ToArray();
+            }
 
             // The data returned is always four channel BGRA
-            var data = ImageReader.Read(stream, out width, out height, out channels, Imaging.STBI_rgb_alpha);
+            var result = ImageResult.FromMemory(bytes, StbImageSharp.ColorComponents.RedGreenBlueAlpha);
 
             // XNA blacks out any pixels with an alpha of zero.
-            if (channels == 4)
+            fixed (byte* b = &result.Data[0])
             {
-                fixed (byte* b = &data[0])
+                for (var i = 0; i < result.Data.Length; i += 4)
                 {
-                    for (var i = 0; i < data.Length; i += 4)
+                    if (b[i + 3] == 0)
                     {
-                        if (b[i + 3] == 0)
-                        {
-                            b[i + 0] = 0;
-                            b[i + 1] = 0;
-                            b[i + 2] = 0;
-                        }
+                        b[i + 0] = 0;
+                        b[i + 1] = 0;
+                        b[i + 2] = 0;
                     }
                 }
             }
 
             Texture2D texture = null;
-            texture = new Texture2D(graphicsDevice, width, height);
-            texture.SetData(data);
+            texture = new Texture2D(graphicsDevice, result.Width, result.Height);
+            texture.SetData(result.Data);
 
             return texture;
         }
